+++ src/main/java/com/andre601/formatterexpansion/FormatterExpansion.java
@@ -4,20 +4,29 @@ import com.andre601.formatterexpansion.formatters.IFormatter;
 import com.andre601.formatterexpansion.formatters.number.NumberFormatter;
 import com.andre601.formatterexpansion.formatters.text.TextFormatter;
 import com.andre601.formatterexpansion.utils.StringUtils;
+import com.andre601.formatterexpansion.utils.logging.LegacyLogger;
+import com.andre601.formatterexpansion.utils.logging.LoggerUtil;
+import com.andre601.formatterexpansion.utils.logging.NativeLogger;
 import me.clip.placeholderapi.PlaceholderAPI;
 import me.clip.placeholderapi.expansion.Configurable;
+import me.clip.placeholderapi.expansion.NMSVersion;
 import me.clip.placeholderapi.expansion.PlaceholderExpansion;
 import org.bukkit.OfflinePlayer;
 
 import javax.annotation.Nonnull;
-import java.util.*;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 
 public class FormatterExpansion extends PlaceholderExpansion implements Configurable{
     
     private final Map<String, Object> defaults = new HashMap<>();
     private final List<IFormatter> formatters;
+    private final LoggerUtil logger;
     
     public FormatterExpansion(){
+        this.logger = loadLogger();
         loadDefaults();
         this.formatters = Arrays.asList(
             new NumberFormatter(this),
@@ -45,6 +54,8 @@ public class FormatterExpansion extends PlaceholderExpansion implements Configur
     
     @Override
     public Map<String, Object> getDefaults(){
+        migrate();
+        
         return defaults;
     }
     
@@ -83,9 +94,22 @@ public class FormatterExpansion extends PlaceholderExpansion implements Configur
         return true;
     }
     
+    public LoggerUtil loadLogger(){
+        if(NMSVersion.getVersion("v1_18_R1") != NMSVersion.UNKNOWN)
+            return new NativeLogger(this);
+        
+        return new LegacyLogger();
+    }
+    
     private void loadDefaults(){
-        defaults.put("format", "#,###,###.##");
-        defaults.put("locale", "en-US");
+        defaults.put("formatting.pattern", "#,###,###.##");
+        defaults.put("formatting.locale", "en-US");
+        
+        defaults.put("shorten.thousands", "K");
+        defaults.put("shorten.millions", "M");
+        defaults.put("shorten.billions", "B");
+        defaults.put("shorten.trillions", "T");
+        defaults.put("shorten.quadrillions", "Q");
     
         defaults.put("time.milliseconds", "ms");
         defaults.put("time.seconds", "s");
@@ -96,5 +120,19 @@ public class FormatterExpansion extends PlaceholderExpansion implements Configur
     
         defaults.put("rounding.precision", 0);
         defaults.put("rounding.mode", "half-up");
+    }
+    
+    private void migrate(){
+        if(this.getString("format", null) == null)
+            return;
+        
+        logger.info("Found old 'format' option. Try to migrate...");
+        
+        defaults.put("format", null);
+        defaults.put("locale", null);
+        defaults.put("formatting.pattern", this.getString("format", "#,###,###.##"));
+        defaults.put("formatting.locale", this.getString("locale", "en-US"));
+        
+        logger.info("Migration completed! Please check the config.yml for any issues.");
     }
 }