+++ src/main/java/com/andre601/formatterexpansion/utils/logging/LegacyLogger.java
@@ -1,0 +1,19 @@
+package com.andre601.formatterexpansion.utils.logging;
+
+import me.clip.placeholderapi.PlaceholderAPIPlugin;
+
+import java.util.logging.Logger;
+
+public class LegacyLogger implements LoggerUtil{
+    
+    private final Logger logger;
+    
+    public LegacyLogger(){
+        logger = PlaceholderAPIPlugin.getInstance().getLogger();
+    }
+    
+    @Override
+    public void info(String msg){
+        logger.info("[Formatter] " + msg);
+    }
+}