+++ src/main/java/com/andre601/formatterexpansion/utils/logging/LoggerUtil.java
@@ -1,0 +1,6 @@
+package com.andre601.formatterexpansion.utils.logging;
+
+public interface LoggerUtil{
+    
+    void info(String msg);
+}